Refactor seems complete
